Re-organize providers.
