+++ docs/googlemaps.rst
@@ -1,4 +1,6 @@
 GoogleMaps
 ----------
 
-.. autojs:: ../lib/providers/googlemaps.js
+.. autojs:: ../lib/providers/googlemaps/provider.js
+
+.. autojs:: ../lib/providers/googlemaps/fixture.js
