+++ docs/openstreetmap.rst
@@ -1,4 +1,6 @@
 OpenStreetMap
 -------------
 
-.. autojs:: ../lib/providers/openstreetmap.js
+.. autojs:: ../lib/providers/openstreetmap/provider.js
+
+.. autojs:: ../lib/providers/openstreetmap/fixture.js
