+++ docs/providers.rst
@@ -3,4 +3,4 @@ Providers
 
 Utilties and base classes to use when writing providers.
 
-.. autojs:: ../lib/providers/provider.js
+.. autojs:: ../lib/providers/utils.js
