+++ lib/index.js
@@ -1,10 +1,7 @@
 var location = require('./location');
-var providers = require('./providers');
 
 module.exports = {
     LocationState: location.LocationState,
-    GoogleMaps: providers.GoogleMaps,
-    OpenStreetMap: providers.OpenStreetMap,
-    providers: providers,
+    providers: require('./providers'),
     test_utils: require('./test_utils'),
 };
