+++ lib/location.js
@@ -5,6 +5,7 @@ var vumigo = require('vumigo_v02');
 var states = vumigo.states;
 var State = states.State;
 var providers = require('./providers');
+var GoogleMaps = providers.googlemaps.GoogleMaps;
 
 
 var LocationState = State.extend(function (self, name, opts) {
@@ -20,6 +21,11 @@ var LocationState = State.extend(function (self, name, opts) {
     :param string_or_LazyText opts.question:
         The question to first display to the user. Defaults to
         ``What is your address?``.
+    :type opts.map_provider:
+        an instance of :class:``location.providers.utils.Provider``
+    :param opts.map_provider:
+        The provider to use when searching for locations. Defaults to
+        an instance of :class:``location.providers.googlemaps.GoogleMaps``.
     :param string_or_LazyText opts.refine_question:
         The question to display to the user when selecting a location from a
         list if the first search query wasn't clear enough. Defaults to
@@ -71,7 +77,7 @@ var LocationState = State.extend(function (self, name, opts) {
             });
         });
 
-    Each provider class providers its own ``.fixture(...)`` function for
+    Each provider has its own ``.fixture(...)`` function for
     conveniently creating HTTP resource fixtures for its location queries. See
     the documentation for each provider on how to use these.
 
@@ -106,7 +112,7 @@ var LocationState = State.extend(function (self, name, opts) {
     self.next_text = opts.next_text;
     self.previous_text = opts.previous_text;
     self.namespace = opts.namespace;
-    self.map_provider = opts.map_provider || new providers.GoogleMaps();
+    self.map_provider = opts.map_provider || new GoogleMaps();
 
     self.init = function() {
         _.defaults(self.metadata, {locationstate:'initial'});
