+++ lib/providers/googlemaps/index.js
@@ -1,0 +1,5 @@
+var provider = require('./provider');
+var fixture = require('./fixture');
+
+this.GoogleMaps = provider.GoogleMaps;
+this.fixture = fixture.fixture;
