+++ lib/providers/index.js
@@ -1,11 +1,3 @@
-var provider = require('./provider');
-var googlemaps = require('./googlemaps');
-var osm = require('./openstreetmap');
-
-this.Provider = provider.Provider;
-this.AddressResult = provider.AddressResult;
-this.ProviderNotImplementedError = provider.ProviderNotImplementedError;
-this.FixtureParameterMissingError = provider.FixtureParameterMissingError;
-
-this.GoogleMaps = googlemaps.GoogleMaps;
-this.OpenStreetMap = osm.OpenStreetMap;
+this.utils = require('./utils');
+this.googlemaps = require('./googlemaps');
+this.openstreetmap = require('./openstreetmap');
