+++ lib/providers/openstreetmap/index.js
@@ -1,0 +1,5 @@
+var provider = require('./provider');
+var fixture = require('./fixture');
+
+this.OpenStreetMap = provider.OpenStreetMap;
+this.fixture = fixture.fixture;
