+++ lib/test_utils.js
@@ -1,7 +1,7 @@
 var assert = require("assert");
 
 var providers = require("./providers");
-var AddressResult = providers.AddressResult;
+var AddressResult = providers.utils.AddressResult;
 
 
 function assert_address_result(result, label, data) {
